move showFolderPicker to Platform, add shared Platform.cpp
