+++ src/OpenLoco/include/OpenLoco/Ui.h
@@ -1,6 +1,5 @@
 #pragma once
 
-#include "OpenLoco/Core/FileSystem.hpp"
 #include <OpenLoco/Engine/Ui/Point.hpp>
 #include <string>
 #include <vector>
@@ -112,7 +111,6 @@ namespace OpenLoco::Ui
     void triggerResize();
     void render();
     void showMessageBox(const std::string& title, const std::string& message);
-    fs::path showFolderPicker(const std::string& title);
     Config::Resolution getResolution();
     Config::Resolution getDesktopResolution();
     bool setDisplayMode(Config::ScreenMode mode, Config::Resolution newResolution);
