+++ src/Platform/src/Platform.Posix.cpp
@@ -7,6 +7,9 @@
 #include <iostream>
 #include <pwd.h>
 #include <time.h>
+#include <SDL3/SDL_dialog.h>
+#include <SDL3/SDL_events.h>
+#include <thread>
 
 #ifdef __linux__
 #include <linux/limits.h>
@@ -104,16 +107,6 @@ namespace OpenLoco::Platform
 #error "Platform does not support full path exe retrieval"
 #endif // __linux__
         return exePath;
-    }
-
-    fs::path promptDirectory([[maybe_unused]] const std::string& Title, [[maybe_unused]] void* hwnd)
-    {
-        std::string input;
-        std::cout << "Type your Locomotion path: ";
-        std::getline(std::cin, input);
-
-        auto path = fs::canonical(input);
-        return path;
     }
 
     std::vector<fs::path> getLocoInstallSearchPaths()
